separate child malnutrition routes
