+++ main.py
@@ -8,7 +8,7 @@ from fastapi.staticfiles import StaticFiles
 from fastapi.responses import HTMLResponse, FileResponse, RedirectResponse
 from Backend.routes.routes import user_router
 from Backend.routes.routes import aaganwadi_summary
-from Backend.routes.routes import child_malnutrition, program_joining
+from Backend.routes.routes import program_joining
 from Backend.routes.ngo_routes import ngo_router
 from Backend.routes.aaganwadi_routes import aanganwadi_router
 from Backend.routes.routes import program_router
@@ -17,6 +17,7 @@ from Backend.routes.child_routes import child_router
 from Backend.routes.donors_routes import donor_router
 from Backend.routes.supplement_packs import supp_router
 from Backend.routes.supplement_details import supplement_details
+from Backend.routes.child_malnutrition_routes import child_malnutrition
 
 # from fastapi.templating import Jinja2Templates
